==> 0_exploratory_database.ipynb <==
import pandas as pd
import os

save_path = "./data/"
read_path = f"{save_path}FoodData_Central_survey_food_csv_2024-10-31/"

csv_files = [f for f in os.listdir(read_path) if f.endswith('.csv')]
csv_files.sort()

dataframes = [pd.read_csv(os.path.join(read_path, key)) for key in csv_files if key.islower()]

# Exploring the data in the `FoodData_Central_survey`

def print_format(dataframes, index):
    print("Dataframe named:", csv_files[index])
    df = dataframes[index]
    return df

print_format(dataframes, 0).head()

food = print_format(dataframes, 1).head()

stripped_food = food[['fdc_id', 'description']]

#Save as CSV
stripped_food.to_csv('./data/PROCESSED_fdc_id_description_pairing.csv', index=False)


print_format(dataframes, 2).head()

food_atribute_ds = print_format(dataframes, 3)

food_atribute_ds['id'].describe()

print_format(dataframes, 5).head()

print_format(dataframes, 6).head()

print_format(dataframes, 7).head()

print_format(dataframes, 8).head()

print_format(dataframes, 10).head()

print_format(dataframes, 11).head()
