==> 2_obtain_nutrient_values_for_iasi_feta.ipynb <==
# 2. Identifying unique aliments in IASI-DATABASE

# Needed Dependencies
# !pip install pandas openpyxl

# Imports
import pandas as pd

For the following FOOD_CODEs ['12012', '17046', '11130', '17045', 'a1362', 'a7021', '12008', '17038', '12042'], we were getting missing values in "./data/FOOD_CODE.csv".

As a result, `FOOD_CODE.csv` has been updated based on [Food Code Lookup](https://www.mrc-epid.cam.ac.uk/wp-content/uploads/2023/03/food_code_lookup.pdf) to include these missing FOOD_CODEs



food_codes_db = pd.read_csv("./data/FOOD_CODES.csv")


food_codes_db['FOOD_CODE'] = food_codes_db['FOOD_CODE'].astype(str).str.strip()
# food_codes_db.iloc[289].FOOD_CODE # DEBUG code

# Get data type for FOOD_CODE
food_codes_db['FOOD_CODE'].dtype

# Verify if the missing FOOD_CODEs are correctly parsed
assert all(code in food_codes_db['FOOD_CODE'].values for code in ['12012', '17046', '11130', '17045', 'a1362', 'a7021', '12008', '17038', '12042'])

food_codes_db["FOOD_CODE"].nunique()

meals_db = pd.read_csv("./data/MEAL_ID.csv")

meals_db.head()

meals_db["LINE NUMBER"].describe()

iasi_db = pd.read_excel('./data/popa2023andreea2013.xlsx')

# Read FOOD_CODE as string to preserve leading 'x' or 'a'
iasi_db['FOOD_CODE'] = iasi_db['FOOD_CODE'].astype(str).str.strip()

# iasi_db['FOOD_CODE'].nunique()
iasi_db

# Group by ID and add FOOD_PORTIONs
iasi_db_grouped = iasi_db.groupby('ID')['FOOD_PORTION'].sum()

iasi_db_grouped[iasi_db_grouped == 320.720000]

# iasi_db_grouped.describe()

iasi_db[iasi_db['FOOD_CODE'] == 'x0535']

unique_vals_b3 = pd.Series(iasi_db['FOOD_CODE'].unique(), name='FOOD_CODE').to_frame()
unique_vals_b3

needed_table = pd.merge(unique_vals_b3, food_codes_db, left_on='FOOD_CODE', right_on='FOOD_CODE', how='left')
# needed_table

# Remove columns that are not needed
# needed_table = needed_table.drop(columns=['AMOUNT', 'MODIFIER'])

# needed_table = pd.merge(needed_table, meals_db, left_on='MEAL', right_on='FFQNAME', how='left')

# Convert to list
# missing_vals = needed_table[needed_table['DESC'].isna()]['FOOD_CODE'].tolist()
# print(missing_vals)

# # Remove columns that are not needed
# needed_table = needed_table.drop(columns=['LINE NUMBER'])

# # Rename FOOD_CODE to FETA_ID
# needed_table = needed_table.rename(columns={'FOOD_CODE': 'FETA_ID'})

# Complete SGDESC by filling NaN with 'Unknown'
# needed_table['SGDESC'] = needed_table['SGDESC'].fillna('Astea apar in baza de date, dar nu ai Index Feta pentru ele in CSVurile tale.')

# needed_table

# Export to CSV
needed_table.to_csv('./data/PROCESSED_needs_fdc_id.csv', index=False)

needed_table.head()

# mini_ds = iasi_db[:10]

# # for fc in mini_ds["FOOD_CODE"]:
# #     print(f"Searching for FOOD_CODE: {fc}")

# #     # Search for FOOD_CODE in food_codes_db
# #     for index, row in food_codes_db.iterrows():
# #         if row["FOOD_CODE"] == fc:
# #             print(f"Found matching FOOD_CODE in food_codes_db: {row['FOOD_CODE']}")
# #             break

# for fc in food_codes_db["FOOD_CODE"]:
#     if fc == 12012:
#         print("Found FOOD_CODE 12012 in food_codes_db")
#         break

# pd.merge(mini_ds, food_codes_db, on="FOOD_CODE", how="left")

# Join iasi_db with food_codes_db on FOOD_CODE
portions_merged_iasi_db = pd.merge(iasi_db, food_codes_db, on='FOOD_CODE', how="left")

portions_merged_iasi_db.head()

# ADDING NUTRIENTS FROM FDC TO IASI-FOODDATA

* A human expert made the manual labelling of the correspodences between FDCs and Feta-FOOD_CODEs in `manually_labelled_fdc_to_feta.xlsx`

# Read xlsx
correspondence_table = pd.read_excel('./data/manually_labelled_fdc_to_feta.xlsx')

correspondence_table['FOOD_CODE'] = correspondence_table['FOOD_CODE'].astype(str).str.strip()

# Combine all fdc_id columns into a single column
correspondence_table['fdc_id'] = correspondence_table[['fdc_id', 'fdc_id.1', 'fdc_id.2', 'fdc_id.3']].apply(lambda x: ','.join(x.dropna().astype(str)), axis=1) 

# Keep all fdc_ids into a single column as list
correspondence_table['fdc_ids'] = correspondence_table.apply(lambda row: [int(float(x)) for x in str(row['fdc_id']).split(',')], axis=1)

# Drop the now redundant columns
correspondence_table = correspondence_table.drop(columns=['fdc_id', 'fdc_id.1', 'fdc_id.2', 'fdc_id.3', 'DESC'])

print(correspondence_table.head(), end='\n\n\n')

print("Statistics of number of fdc_ids per FOOD_CODE:", end='\n')
print(correspondence_table['fdc_ids'].apply(lambda x: len(x)).describe())

partial_merged_df = pd.merge(portions_merged_iasi_db, correspondence_table, on='FOOD_CODE', how='left')
partial_merged_df.head()

nutrients_of_interest = pd.read_csv('./data/PROCESSED_filtered_nutrients_multihot.csv')

nutrients_of_interest.rename(columns={'description': 'double_check_desc'}, inplace=True)

nutrients_of_interest.head()

nutrient_columns = nutrients_of_interest.columns.tolist()
nutrient_columns.remove('fdc_id')

raw_fully_merged_df = partial_merged_df.copy()

for index, row in partial_merged_df.iterrows():
    fdc_ids = row['fdc_ids']
    if isinstance(fdc_ids, list):
        nutrient_values = {}

        # If only one fdc_id, directly get the nutrient values
        if len(fdc_ids) == 1:
            fdc_id = fdc_ids[0]
            nutrient_row = nutrients_of_interest[nutrients_of_interest['fdc_id'] == fdc_id]
            if not nutrient_row.empty:
                for col in nutrient_columns:
                    raw_fully_merged_df.at[index, col] = nutrient_row.iloc[0][col]
        else:
            # Multiple fdc_ids, compute average for each nutrient
            # For the description, append all descriptions together

            # Initialize sums and counts with the first fdc_id
            nutrient_sums = nutrients_of_interest[nutrients_of_interest['fdc_id'] == fdc_ids[0]][nutrient_columns].copy()
            counts = 1

            # print(nutrient_sums)

            # Add values from the remaining fdc_ids
            for fdc_id in fdc_ids[1:]:
                nutrient_row = nutrients_of_interest[nutrients_of_interest['fdc_id'] == fdc_id]
                if not nutrient_row.empty:
                    counts += 1
                    for col in nutrient_columns:
                        value = nutrient_row.iloc[0][col]
                        if pd.notna(value):
                            # Append description or sum nutrients
                            if col == 'double_check_desc':
                                if pd.isna(raw_fully_merged_df.at[index, col]):
                                    raw_fully_merged_df.at[index, col] = value
                                else:
                                    raw_fully_merged_df.at[index, col] += f"; {value}"
                            else:
                                nutrient_sums[col] += float(value)
            # Compute averages for nutrients
            for col in nutrient_columns:
                if col != 'double_check_desc':
                    if counts > 0:
                        # Choosed to use iloc[0] for future compatibilitysince nutrient_sums[col] returns a "Deprecated" warning
                        raw_fully_merged_df.at[index, col] = nutrient_sums.iloc[0][col] / counts

                    else:
                        raw_fully_merged_df.at[index, col] = pd.NA

raw_fully_merged_df.head()


raw_fully_merged_df

raw_fully_merged_df.to_csv('./data/PROCESSED_iasi_with_nutrients_raw.csv', index=False)
