==> 3_iasi_results.ipynb <==
# 3. REFINED RESULTS AND STATISTICS

In the USDA database, it is assumed that the nutrient values are per 100g of the food item.
In IASI database, FOOD_PORTION is given in grams.
Therefore, we need to scale the nutrient values accordingly:

$ \text{Nutrient Value (scaled)} = \left( \frac{\text{Nutrient Value (per 100g)} \times \text{FOOD\_PORTION (g)}}{100} \right) $

# Imports
import pandas as pd

refined_iasi_df = pd.read_csv("./data/PROCESSED_iasi_with_nutrients_raw.csv")
refined_iasi_df.head()

# Drop columns food_code, fdc_ids, double_check_desc
refined_iasi_df = refined_iasi_df.drop(columns=['FOOD_CODE', 'fdc_ids', 'double_check_desc'])
refined_iasi_df.head()

nutrient_columns = refined_iasi_df.columns.difference(['ID', 'MEAL_ID', 'FOOD_PORTION', 'DESC'])

# In the USDA database, it is assumed that the nutrient values are per 100g of the food item.
# In IASI database, FOOD_PORTION is given in grams.
# Therefore, we need to scale the nutrient values accordingly.
for col in nutrient_columns:
    refined_iasi_df[col] = (refined_iasi_df[col] * refined_iasi_df['FOOD_PORTION']) / 100.0

refined_iasi_df.to_csv('./data/PROCESSED_iasi_with_nutrients_scaled_with_food_portion.csv', index=False)
refined_iasi_df.head()

refined_iasi_per_patient_db = refined_iasi_df.groupby('ID')[nutrient_columns].sum().reset_index()

refined_iasi_per_patient_db.to_csv('./data/PROCESSED_iasi_per_patient_nutrient_totals.csv', index=False)

refined_iasi_per_patient_db

### PLOTS AND STATISTICS - TBD
